% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{attach_to_study}
\alias{attach_to_study}
\title{Attach Data API to a study}
\usage{
attach_to_study(
  db_id,
  table_id,
  idno,
  dataset_title,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(required) DB ID}

\item{table_id}{(required) Table ID}

\item{idno}{(required) Study unique identifier}

\item{dataset_title}{(required) Dataset title}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\description{
Attach dataset available via API to a study
}
\examples{

attach_to_study (
  db_id="example",
  table_id="prices",
  idno="survey-idno-test",
  dataset_title = "Dataset title"
)

}
