% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_pdf_cover}
\alias{capture_pdf_cover}
\title{Capture cover page of a PDF}
\usage{
capture_pdf_cover(file_path, dpi = 72)
}
\arguments{
\item{file_path}{\strong{(required)} PDF file path}

\item{dpi}{set image DPI, default is 72 dpi}
}
\value{
image file
}
\description{
Capture cover page of a PDF file
}
