% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R
\name{catalog_delete}
\alias{catalog_delete}
\title{Delete a catalog study}
\usage{
catalog_delete(idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(Required) Dataset IDNo}
}
\value{
list
}
\description{
Delete a single entry from the catalog
}
