% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R
\name{catalog_list}
\alias{catalog_list}
\title{List catalog entries}
\usage{
catalog_list(idno = NULL, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(Optional) Dataset IDNo}
}
\value{
List of studies or a single study info
}
\description{
List all entries in the catalog
}
