% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collection_add}
\alias{collection_add}
\title{Create new collection}
\usage{
collection_add(
  repositoryid = NULL,
  title = NULL,
  short_text = NULL,
  long_text = NULL,
  thumbnail = NULL,
  weight = 0,
  section = 0,
  ispublished = 0,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{repositoryid}{(Required) new collection ID}

\item{title}{Title of the collection}

\item{short_text}{Short description of the collection}

\item{long_text}{Long description of the collection}

\item{thumbnail}{Thumbnail file path}

\item{weight}{Weight}

\item{section}{Section}

\item{ispublished}{Set collection status - 1 = Published, 0 = Unpublished}
}
\description{
Create new collection
}
