% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collection_delete}
\alias{collection_delete}
\title{Delete a collection}
\usage{
collection_delete(collection_idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{collection_idno}{(Required) Collection IDNo}
}
\value{
status
}
\description{
Delete a collection
}
