% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection_rename}
\alias{collection_rename}
\title{Rename collection}
\usage{
collection_rename(old_idno, new_idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{old_idno}{(Required) old collection IDNO}

\item{new_idno}{(Required) new collection IDNO}
}
\description{
Rename collection
}
