% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collections}
\alias{collections}
\title{Get all collections}
\usage{
collections(collection_idno = NULL, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{collection_idno}{(Optional) Collection IDNo}
}
\value{
List of studies or a single study info
}
\description{
Load a list of all collections or get info for a single collection
}
