% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{create_collection}
\alias{create_collection}
\title{Create new collection}
\usage{
create_collection(
  collectionid,
  title,
  short_text,
  long_text,
  thumbnail = NULL,
  weight = 0,
  section = 2,
  ispublished = 0,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{collectionid}{(required) Collection identifier containing numbers and letters only}

\item{title}{(required) Collection title}

\item{short_text}{(required) A short description for the collection}

\item{long_text}{(required) Detailed collection description. This field supports basic html and image tags.}

\item{thumbnail}{Thumbnail image - provide path/url}

\item{weight}{Provide weight to arrange display of collection}

\item{section}{(required) Section/Group. To see a list of sections available, see api endpoint collections/sections}

\item{ispublished}{(required) Set status to draft or published}
}
\description{
Create new collection
}
