% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_entry.R
\name{create_entry}
\alias{create_entry}
\title{Create entry in catalog}
\usage{
create_entry(
  type,
  idno,
  metadata,
  repositoryid = NULL,
  access_policy = NULL,
  data_remote_url = NULL,
  published = 0,
  overwrite = "no",
  thumbnail = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{type}{\strong{(required)} Entry type - "document", "script", "table",}

\item{idno}{\strong{(required)} Study unique identifier}

\item{metadata}{\strong{(required)} Metadata list depending on the type of study}

\item{repositoryid}{Collection ID that owns the entry}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open",
"direct", "public", "licensed", "enclave", "remote", "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for entry - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a entry with the same ID already exists? Valid values "yes", "no"}

\item{api_key}{(optoinal) API key}

\item{api_base_url}{(optional) API URL}
}
\value{
result
}
\description{
Create entry in catalog
}
