% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{create_geospatial}
\alias{create_geospatial}
\title{Create new geospatial study}
\usage{
create_geospatial(
  api_key = NULL,
  api_base_url = NULL,
  idno,
  repositoryid = NULL,
  access_policy = NULL,
  data_remote_url = NULL,
  published = NULL,
  overwrite = NULL,
  metadata = NULL
)
}
\arguments{
\item{idno}{(required) Study unique identifier}

\item{repositoryid}{Collection ID that owns the study}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}

\item{metadata}{\strong{(required)} Metadata list depending on the type of study}
}
\description{
Create a new geospatial study
}
\examples{

metadata= list(
metadata_maintenance=list(
"update_frequency"="daily",
"note"="Maintenane note",
"contact"=list(
list(
    "person_name"="name here",
    "role"="role"
  )
)
)

dataset_description=list(
"file_identifier"="Unique IDNO",
"language"="en",
"contact"= list(
    "person_name"= "name",
    "role"= "string",
  ),
  "identification_info"=list(
     "title"="title goes here",
  )
)
)

create_geospatial (
  idno="survey-idno-test",
  published = 1,
  overwrite = "yes",
  metadata_maintenance = list(),
  dataset_description = list()
  additional = list()
)




}
