% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{create_resource}
\alias{create_resource}
\title{Create new resource}
\usage{
create_resource(idno, dctype, title, dcformat = NULL, author = NULL,
  dcdate = NULL, country = NULL, language = NULL,
  contributor = NULL, publisher = NULL, rights = NULL,
  description = NULL, abstract = NULL, toc = NULL,
  file_path = NULL, overwrite = "no", api_key = NULL,
  api_base_url = NULL)
}
\arguments{
\item{idno}{\strong{(required)} Study IDNO}

\item{dctype}{Resource document type}

\item{title}{Resource title}

\item{dcformat}{Resource file format}

\item{author}{Author name}

\item{dcdate}{Date using YYYY-MM-DD format}

\item{country}{Country name}

\item{language}{Language or Language code}

\item{contributor}{Contributor name}

\item{publisher}{Publisher name}

\item{rights}{Rights}

\item{description}{Resource detailed description}

\item{abstract}{Resource abstract}

\item{toc}{Table of contents}

\item{file_path}{File path for uploading}

\item{overwrite}{Overwrite if resource already exists? Accepted values "yes", "no"}
}
\description{
Create a new resource
}
