% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{create_survey}
\alias{create_survey}
\title{Create Survey}
\usage{
create_survey(idno, repositoryid = NULL, access_policy = NULL,
  data_remote_url = NULL, published = NULL, overwrite = NULL,
  doc_desc, study_desc, data_files = NULL, variables = NULL,
  variable_groups = NULL, additional = NULL, api_key = NULL,
  api_base_url = NULL)
}
\arguments{
\item{idno}{(required) Survey unique identifier}

\item{repositoryid}{Collection ID that owns the study}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}

\item{doc_desc}{Document description}

\item{study_desc}{(Required) Study description}

\item{data_files}{Data files}

\item{variables}{List of variables and metadata}

\item{variable_groups}{List of variable groups}

\item{additional}{Additional metadata}

\item{type}{(required) DDI/XML file path}
}
\description{
Create a new survey
}
\examples{
doc_desc=list(
"idno"="doc-idno",
"producers"=list(
list(
    "name"="name here",
    "abbr"="abbreviation"
  )
)
)

study_desc=list(
  "title_statement"= list(
    "idno"= "survey-idno-test",
    "title"= "string",
    "sub_title"= "string",
    "alternate_title"= "string",
    "translated_title"= "string"
  ),
  "study_info"=list(
    "nation"=list(
      list(
        "name"="Test",
        "abbreviation"="tst")
    )
  )
)

createSurvey (
  idno="survey-idno-test",
  type="survey",
  published = 1,
  overwrite = "yes",
  doc_desc = doc_desc,
  study_desc = study_desc,
  data_files=NULL,
  variables=NULL,
  variable_groups = NULL,
  additional = NULL
)




}
