% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_api_create_table}
\alias{data_api_create_table}
\title{createTable}
\usage{
data_api_create_table(
  db_id,
  table_id,
  metadata,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{metadata}{Table metadata}
}
\description{
Create a table
}
