% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_api_delete_table}
\alias{data_api_delete_table}
\title{deleteTable}
\usage{
data_api_delete_table(db_id, table_id, api_key = NULL,
  api_base_url = NULL)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}
}
\description{
delete a table
}
