% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_api_publish_table}
\alias{data_api_publish_table}
\title{publishTable}
\usage{
data_api_publish_table(
  db_id,
  table_id,
  table_metadata,
  csvfile,
  overwrite = "no",
  api_key = NULL,
  api_base_url = NULL,
  sleep_seconds = 1
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{metadata}{Table metadata}
}
\description{
Create a data table with data using CSV
}
