% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_api.R
\name{data_api_timeseries_create}
\alias{data_api_timeseries_create}
\title{dataApiCreateTimeseries}
\usage{
data_api_timeseries_create(
  db_id,
  table_id,
  metadata,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{metadata}{Table metadata}
}
\description{
Data API - create timeseries table
}
