% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_api.R
\name{data_api_timeseries_import_csv}
\alias{data_api_timeseries_import_csv}
\title{dataApiTimeseriesImportCSV}
\usage{
data_api_timeseries_import_csv(
  db_id,
  table_id,
  csvfile,
  overwrite = "no",
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{csvfile}{CSV file}

\item{overwrite}{(valid values = "yes" or "no")}
}
\description{
Data API - Timeseries import CSV
}
