% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_api.R
\name{data_api_timeseries_publish}
\alias{data_api_timeseries_publish}
\title{publishTimeseriesTable}
\usage{
data_api_timeseries_publish(
  db_id,
  table_id,
  table_metadata,
  csvfile,
  overwrite = "no",
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{metadata}{Table metadata}
}
\description{
Data API - publish timeseries table
}
