% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{datafile_data_dictionary}
\alias{datafile_data_dictionary}
\title{Generate data dictionary (STATA, SPSS, CSV)}
\usage{
datafile_data_dictionary(filepath, type = "CSV", fileId = "F1", freqLimit = 50)
}
\arguments{
\item{type}{Type of data file (dta | sav | csv)}

\item{fileId}{unique file id e.g. F1, F2, ...}

\item{freqLimit}{frequency limit}

\item{file}{Path to data file or upload file}
}
\value{
data dictionary list
}
\description{
Generate data dictionary from various file formats (STATA, SPSS, CSV)
}
\examples{
 import('\\\\Sample Data\\\\in\\\\HND_2012_L2L.dta', 'dta', 'F1', 36000)

}
