% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{dataset_attach_collections}
\alias{dataset_attach_collections}
\title{Set owner and linked collections for studies}
\usage{
dataset_attach_collections(
  study_idno,
  owner_collection = NULL,
  link_collections = list(),
  mode = "update",
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{study_idno}{(Required) Study/dataset IDNo}

\item{owner_collection}{(Optional) Owner Collection IDNo}

\item{link_collections}{List of collections to add to a study/dataset}

\item{mode}{(Required) Update or replace the linked collections. Values 'update' or 'replace'. Default is 'update'}
}
\description{
Set owner and linked collections for studies
}
