% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_options}
\alias{dataset_options}
\title{Set various options for dataset}
\usage{
dataset_options(
  idno,
  api_key = NULL,
  api_base_url = NULL,
  access_policy = NULL,
  data_remote_url = NULL,
  published = NULL,
  tags = NULL,
  aliases = NULL,
  owner_collection = NULL,
  linked_collections = NULL,
  link_study = NULL,
  link_indicator = NULL
)
}
\arguments{
\item{idno}{(required) Study unique identifier}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open", "direct", "public", "licensed", "remote" ("enclave", "other")}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to "remote".}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{tags}{Tag or vector of multiple tags for study (string)}

\item{aliases}{Alias or vector of multiple aliases for study (string)}

\item{owner_collection}{Collection that owns the dataset (repositoryid (string) of existing collection)}

\item{linked_collections}{Display in other collections (repositoryid (string) of existing collection or vector with multiple collections)}

\item{link_study}{URL for study website (string must include http:// or https://)}

\item{link_indicator}{URL to the indicators website (string must include http:// or https://)}

\item{verbose}{Show verbose output - True, False}
}
\description{
Set various options for dataset, such as access policy, project publish status, tags and aliases, owner and linked collections and links to the data, study website and indicators website.
}
\examples{

dataset_options (
  idno="survey-idno-test",
  access_policy = "licensed",
  tags = "ihsn",
  link_study = "http://www.studypage.org",
  link_indicator = "http://www.indicatorpage.org"
)

}
