% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{datasetExternalResources}
\alias{datasetExternalResources}
\title{List External Resources}
\usage{
datasetExternalResources(api_key = NULL, api_base_url = NULL, idno)
}
\arguments{
\item{idno}{Study IDNo}
}
\value{
List of external resources
}
\description{
List external resources for a study
}
