% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{GetDatasets}
\usage{
datasets(api_key = NULL, api_base_url = NULL, idno = NULL)
}
\arguments{
\item{idno}{(Optional) Dataset IDNo}
}
\value{
List of studies or a single study info
}
\description{
Load a list of all datasets or get info for a single dataset/study
}
