% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{delete_entry}
\alias{delete_entry}
\title{Delete a study}
\usage{
delete_entry(idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(Required) Dataset IDNo}
}
\value{
list
}
\description{
Delete a single entry from the catalog
}
