% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_compability.R
\name{document_add}
\alias{document_add}
\title{(deprecated - use add_document) Create new document}
\usage{
document_add(idno, metadata, repositoryid = NULL, access_policy = NULL,
  data_remote_url = NULL, published = 0, overwrite = "no",
  thumbnail = NULL, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{\strong{(required)} Entry unique identifier}

\item{metadata}{\strong{(required)} Metadata list}

\item{repositoryid}{Collection ID that owns the study}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for entry - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a entry with the same ID already exists? Valid values "yes", "no"}
}
\description{
Deprecated function, available for backward compability
}
\details{
Create a new document
}
