% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.calculate_catFrequency}
\alias{.calculate_catFrequency}
\title{calculate category frequency}
\usage{
.calculate_catFrequency(df, catgryDF, varName, ...)
}
\arguments{
\item{df}{dataframe}

\item{catgryDF}{categories}

\item{varName}{variable name}
}
\value{
list of categories
}
\description{
calcuilate category frequency od the selected variable
}
\examples{
 .calculate_catFrequency(x)

}
