% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.format_num}
\alias{.format_num}
\title{format number}
\usage{
.format_num(x, dcml = 0, ...)
}
\arguments{
\item{x}{variable}

\item{dcml}{number of decimal}
}
\value{
formatted variable
}
\description{
convert numeric vector to string format with correct decimal precisions
}
\examples{
 .format_num(x)

}
