% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{.nada_resolve_resource_url}
\alias{.nada_resolve_resource_url}
\title{Build full download URL from public resources response}
\usage{
.nada_resolve_resource_url(url, api_base_url = NULL)
}
\description{
Build full download URL from public resources response
}
\keyword{internal}
