% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{.nada_resource_row_by_id}
\alias{.nada_resource_row_by_id}
\title{Find resource row by resource_id in list response}
\usage{
.nada_resource_row_by_id(resources, resource_id)
}
\description{
Find resource row by resource_id in list response
}
\keyword{internal}
