% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.read_matchingVariables}
\alias{.read_matchingVariables}
\title{read matching variables}
\usage{
.read_matchingVariables(variables, file = "", default = "?", ...)
}
\arguments{
\item{variables}{variables}

\item{file}{data file}

\item{default}{default type to be used when reading the CSV file}
}
\value{
dataframe
}
\description{
Read CSV file and return only the matching variables
}
\examples{
.read_matchingVariables(var1, 'c://f1.csv' )

}
