% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.stata_datatype}
\alias{.stata_datatype}
\title{find stata data type}
\usage{
.stata_datatype(data, ...)
}
\arguments{
\item{data}{variable}
}
\value{
stata data type
}
\description{
find stata data type of a numeric variable
}
\examples{
.stata_datatype(data)

}
