% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.summary_stats}
\alias{.summary_stats}
\title{variable statistics}
\usage{
.summary_stats(df, varName, ...)
}
\arguments{
\item{df}{dataframe}

\item{varName}{variable name}
}
\value{
A list contain variable statistics
}
\description{
calculate variable statistics
}
\examples{
 .summary_stats(df,'uqnr')

}
