% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.varibale_width}
\alias{.varibale_width}
\title{variable width}
\usage{
.varibale_width(variable, type, catValues, ...)
}
\arguments{
\item{variable}{variable}

\item{type}{variable type}

\item{catValues}{variable categories}
}
\value{
variable width
}
\description{
Calculate the variable width
}
\examples{
.varibale_width(var1, 'numeric', c() )

}
