% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.varStats}
\alias{.varStats}
\title{variable summary}
\usage{
.varStats(df, varName, ...)
}
\arguments{
\item{df}{dataframe}

\item{varName}{variable name}
}
\value{
A list contain variable summary statistics
}
\description{
get variable summary statistics
}
\examples{
 .varStats(df,'uqnr')

}
