% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_utils.R
\name{.write_fwf}
\alias{.write_fwf}
\title{Write fixed width format}
\usage{
.write_fwf(
  x,
  file = "",
  locale = "",
  encoding = "UTF-8",
  append = FALSE,
  quote = FALSE,
  sep = " ",
  na = "",
  rownames = FALSE,
  colnames = TRUE,
  rowCol = NULL,
  justify = "left",
  formatInfo = FALSE,
  quoteInfo = TRUE,
  width = NULL,
  eol = "\\n",
  qmethod = c("escape", "double"),
  scientific = TRUE,
  ...
)
}
\arguments{
\item{x}{dataframe}

\item{file}{output file path}
}
\value{
write format
}
\description{
Write fixed width format (Note: the write.fwf method is copied from 'gdata' R package since we are not able to load the package because of the dependencies like perl)
}
\examples{
.write_fwf(df, 'c://test.txt')

}
