% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_project.R
\name{editor_create}
\alias{editor_create}
\title{Create new project}
\usage{
editor_create(
  type,
  metadata,
  idno = NULL,
  thumbnail = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{type}{(required) Type of project - survey, geospatial, table, document, timeseries}

\item{metadata}{\strong{(required)} Metadata list depending on the type of study}

\item{idno}{\strong{(required)} Project unique IDNO}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\description{
Create a new project
}
