% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_resources.R
\name{editor_resources_delete}
\alias{editor_resources_delete}
\title{Editor - delete External Resources}
\usage{
editor_resources_delete(idno, resource_id, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{Project IDNo}

\item{resource_id}{Resource ID}
}
\description{
Delete external resources for a project
}
