% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_resources.R
\name{editor_resources_import_rdf}
\alias{editor_resources_import_rdf}
\title{Editor import RDF}
\usage{
editor_resources_import_rdf(
  idno,
  rdf_file,
  skip_uploads = FALSE,
  overwrite = "no",
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{Project IDNo}

\item{rdf_file}{RDF file path}

\item{skip_uploads}{TRUE/FALSE - If TRUE, won't upload files}

\item{overwrite}{yes/no - Overwrite existing resources}
}
\description{
Import an RDF file
}
