% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_resources.R
\name{editor_resources_upload}
\alias{editor_resources_upload}
\title{Editor upload external resource file}
\usage{
editor_resources_upload(idno, file, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{Project IDNo}

\item{file}{External resource file to be uploaded}
}
\description{
Upload an external resource file
}
