% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{external_resources_delete}
\alias{external_resources_delete}
\title{Delete External Resources}
\usage{
external_resources_delete(
  dataset_idno,
  resource_id,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{resource_id}{Resource ID}
}
\value{

}
\description{
Delete external resources for a study
}
