% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{external_resources_delete_all}
\alias{external_resources_delete_all}
\title{Delete all external resources for a study}
\usage{
external_resources_delete_all(
  dataset_idno,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}
}
\description{
Delete all external resources for a study
}
