% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{external_resources_import}
\alias{external_resources_import}
\title{import_rdf}
\usage{
external_resources_import(
  dataset_idno,
  rdf_file,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{rdf_file}{RDF file path}
}
\description{
Import an RDF file
}
