% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{external_resources_list}
\alias{external_resources_list}
\title{List External Resources}
\usage{
external_resources_list(dataset_idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{dataset_idno}{Study IDNo}
}
\value{
List of external resources
}
\description{
List external resources for a study
}
