% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{external_resources_upload}
\alias{external_resources_upload}
\title{Upload external resources}
\usage{
external_resources_upload(
  dataset_idno,
  resource_id = NULL,
  file,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{resource_id}{(Optional) External resource ID}

\item{file}{External resource file to be uploaded}
}
\description{
Upload an external resource file
}
