% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pdf_coverpage}
\alias{generate_pdf_coverpage}
\title{Generate PDF cover page thumbnail}
\usage{
generate_pdf_coverpage(file_path, dpi = 72)
}
\arguments{
\item{file_path}{\strong{(required)} PDF file path}

\item{dpi}{set image DPI, default is 72 dpi}
}
\value{
image file
}
\description{
Generate PDF cover page thumbnail
}
