% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{geospatial_import}
\alias{geospatial_import}
\title{Import geospatial xml file}
\usage{
geospatial_import(file_uri, repositoryid = NULL, access_policy = NULL,
  data_remote_url = NULL, published = 0, overwrite = "no",
  api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{repositoryid}{Collection ID that owns the study}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}

\item{file}{\strong{(required)} File or URL of ISO19139 XML file}
}
\description{
Import geospatial xml (ISO19139)
}
\examples{


geospatial_import (
  published = 1,
  overwrite = "yes",
  file = "http://example.com/file/file-iso13913.xml",
)




}
