% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R, R/globals.R
\name{get_api_key}
\alias{get_api_key}
\title{Get API key}
\usage{
get_api_key()

get_api_key()
}
\value{
API key generated in NADA

API Key
}
\description{
`get_api_key` returns the API key from global environment variables.

Get API key from global Environment variable
}
