% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R, R/globals.R
\name{get_api_url}
\alias{get_api_url}
\title{Get API URL}
\usage{
get_api_url(endpoint = NULL)

get_api_url(endpoint = NULL)
}
\arguments{
\item{endpoint}{Endpoint to be added to the base endpoint URL, e.g., ""}
}
\value{
API

API base url
}
\description{
`get_api_key` returns the API base endpoint URL from global environment variables.

Get API base endpoint
}
