% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_api_verbose}
\alias{get_api_verbose}
\title{get_verbose}
\usage{
get_api_verbose()
}
\value{
verbose parameter value
}
\description{
`get_api_verbose`  returns the API verbose parameter from the global environment variables.
}
