% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection_info}
\alias{get_collection_info}
\title{Returns information on a collection}
\usage{
get_collection_info(repository_id, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{repository_id}{Collection's unique IDNo}
}
\description{
Returns information on a collection
}
