% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_test_db}
\alias{get_config_test_db}
\title{Get test database name from config or environment}
\usage{
get_config_test_db()
}
\value{
database name string
}
\description{
Get test database name from config or environment
}
