% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_test_table}
\alias{get_config_test_table}
\title{Get test table name from config or environment}
\usage{
get_config_test_table()
}
\value{
table name string
}
\description{
Get test table name from config or environment
}
