% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_api.R
\name{get_editor_api_key}
\alias{get_editor_api_key}
\title{Get Editor API key}
\usage{
get_editor_api_key()
}
\value{
API key generated in NADA
}
\description{
`get_editor_api_key` returns the API key from global environment variables.
}
