% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_api.R
\name{get_editor_api_url}
\alias{get_editor_api_url}
\title{Get Editor API URL}
\usage{
get_editor_api_url(endpoint = NULL)
}
\arguments{
\item{endpoint}{Endpoint to be added to the base endpoint URL, e.g., ""}
}
\value{
API
}
\description{
`get_editor_api_key` returns the API base endpoint URL from global environment variables.
}
