% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_api.R
\name{get_editor_api_verbose}
\alias{get_editor_api_verbose}
\title{get_editor_verbose}
\usage{
get_editor_api_verbose()
}
\value{
verbose parameter value
}
\description{
`get_api_verbose`  returns the API verbose parameter from the global environment variables.
}
