% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pdf_coverpage}
\alias{get_pdf_coverpage}
\title{Get PDF cover page thumbnail}
\usage{
get_pdf_coverpage(file_path, dpi = 72)
}
\arguments{
\item{file_path}{\strong{(required)} PDF file path}

\item{dpi}{set image DPI, default is 72 dpi}
}
\value{
image file
}
\description{
Get PDF cover page thumbnail
}
