% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{has_real_api_config}
\alias{has_real_api_config}
\title{Check if real API configuration is available}
\usage{
has_real_api_config()
}
\value{
logical indicating if configuration is available
}
\description{
Check if real API configuration is available
}
