% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{importRDF}
\alias{importRDF}
\title{ImportRDF}
\usage{
importRDF(api_key = NULL, api_base_url = NULL, dataset_idno, rdf_file)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{rdf_file}{RDF file path}
}
\description{
Import an RDF file
}
