% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{list_collections}
\alias{list_collections}
\title{Returns a list of all collections in the catalog}
\usage{
list_collections(api_key = NULL, api_base_url = NULL)
}
\description{
Returns a list of all collections in the catalog
}
