% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{nada_admin_codelist_delete}
\alias{nada_admin_codelist_delete}
\title{Delete a codelist}
\usage{
nada_admin_codelist_delete(id, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{id}{Numeric id.}

\item{api_key, api_base_url}{See `nada_admin_dsd_list`.}
}
\value{
list with status_code and response
}
\description{
Endpoint: DELETE /api/admin/codelists/item/{id}
}
