% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{nada_admin_codelist_update}
\alias{nada_admin_codelist_update}
\title{Update a codelist}
\usage{
nada_admin_codelist_update(id, fields, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{id}{Numeric id.}

\item{fields}{Named list of fields to update (name, idno, agency, version, description).}

\item{api_key, api_base_url}{See `nada_admin_dsd_list`.}
}
\value{
list with status_code and response
}
\description{
Endpoint: PUT /api/admin/codelists/item/{id}
}
