% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{nada_admin_collection_delete}
\alias{nada_admin_collection_delete}
\title{Delete a collection}
\usage{
nada_admin_collection_delete(repositoryid, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{repositoryid}{(Required) Collection IDNo}
}
\value{
status
}
\description{
Delete a collection
}
