% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{nada_admin_collection_update}
\alias{nada_admin_collection_update}
\title{update collection}
\usage{
nada_admin_collection_update(
  repositoryid = NULL,
  title = NULL,
  short_text = NULL,
  long_text = NULL,
  thumbnail = NULL,
  weight = NULL,
  section = NULL,
  ispublished = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{repositoryid}{(Required) Collection ID to be updated}

\item{title}{Title of the collection}

\item{short_text}{Short description of the collection}

\item{long_text}{Long description of the collection}

\item{thumbnail}{Thumbnail file path}

\item{weight}{Weight}

\item{section}{Section}

\item{ispublished}{Set collection status - 1 = Published, 0 = Unpublished}
}
\description{
Update collection options - all fields are optional
}
