% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nada_admin_data_table_batch_import_csv}
\alias{nada_admin_data_table_batch_import_csv}
\title{Batch Import CSV}
\usage{
nada_admin_data_table_batch_import_csv(
  db_id,
  table_id,
  max_rows = NULL,
  delimiter = ",",
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{max_rows}{Number of rows to be processed in a batch}

\item{delimiter}{CSV delimiter: comma, tab, semi-colon, colon (default: "comma")}

\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}
}
\description{
Import uploaded CSV or ZIP file into the database using the updated API endpoint
}
\examples{

nada_admin_data_table_batch_import_csv (
  db_id = "example",
  table_id = "prices"
)

nada_admin_data_table_batch_import_csv (
  db_id = "example",
  table_id = "prices",
  max_rows = 5000,
  delimiter = "semi-colon"
)

}
