% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nada_admin_data_table_create}
\alias{nada_admin_data_table_create}
\title{createTable}
\usage{
nada_admin_data_table_create(
  db_id,
  table_id,
  metadata,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{metadata}{Table metadata}
}
\description{
Create a table
}
