% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nada_admin_data_table_delete}
\alias{nada_admin_data_table_delete}
\title{deleteTable}
\usage{
nada_admin_data_table_delete(
  db_id,
  table_id,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}
}
\description{
delete a table
}
