% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nada_admin_data_table_import_csv}
\alias{nada_admin_data_table_import_csv}
\title{dataApiImportCSV}
\usage{
nada_admin_data_table_import_csv(
  db_id,
  table_id,
  max_rows = NULL,
  delimiter = ",",
  api_key = NULL,
  api_base_url = NULL,
  max_batches = 1000,
  sleep_seconds = 1
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{max_rows}{Number of rows to be processed in a batch}

\item{delimiter}{CSV delimiter: comma, tab, semi-colon, colon (default: ",")}

\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}

\item{max_batches}{Maximum number of batches to process (safety limit)}

\item{sleep_seconds}{Sleep time in seconds between batch calls (default: 1)}
}
\description{
Import CSV using batch import process
}
