% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{nada_admin_dsd_delete}
\alias{nada_admin_dsd_delete}
\title{Delete a data structure}
\usage{
nada_admin_dsd_delete(id_or_idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{id_or_idno}{Numeric id or catalogue idno of the data structure.}

\item{api_key, api_base_url}{See `nada_admin_dsd_list`.}
}
\value{
list with status_code and response
}
\description{
Endpoint: POST /api/admin/data_structures/delete/{id_or_idno}
}
