% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\name{nada_admin_microdata_add}
\alias{nada_admin_microdata_add}
\title{Create Microdata}
\usage{
nada_admin_microdata_add(
  idno,
  repositoryid = "central",
  access_policy = NULL,
  data_remote_url = NULL,
  published = 1,
  overwrite = "no",
  metadata = NULL,
  thumbnail = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{(required) Survey unique identifier}

\item{repositoryid}{Collection ID that owns the study}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}

\item{type}{(required) DDI/XML file path}

\item{doc_desc}{Document description}

\item{study_desc}{(Required) Study description}

\item{data_files}{Data files}

\item{variables}{List of variables and metadata}

\item{variable_groups}{List of variable groups}

\item{additional}{Additional metadata}
}
\description{
Create a new Microdata study
}
\examples{

metadata=list(
doc_desc=list(
"idno"="doc-idno",
"producers"=list(
list(
    "name"="name here",
    "abbr"="abbreviation"
  )
)
),
study_desc=list(
  "title_statement"= list(
    "idno"= "survey-idno-test",
    "title"= "string",
    "sub_title"= "string",
    "alternate_title"= "string",
    "translated_title"= "string"
  ),
  "study_info"=list(
    "nation"=list(
      list(
        "name"="Test",
        "abbreviation"="tst")
    )
  )
 )
)

nada_admin_microdata_add (
  idno="survey-idno-test",
  type="survey",
  published = 1,
  overwrite = "yes",
  metadata = metadata,
  additional = NULL
)




}
