% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{nada_admin_resource_delete}
\alias{nada_admin_resource_delete}
\title{Delete External Resources}
\usage{
nada_admin_resource_delete(
  dataset_idno,
  resource_id,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{resource_id}{Resource ID}
}
\value{
List with status_code and response
}
\description{
Delete external resources for a study
}
