% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{nada_admin_resource_delete_all}
\alias{nada_admin_resource_delete_all}
\title{Delete all external resources for a study}
\usage{
nada_admin_resource_delete_all(
  dataset_idno,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}
}
\value{
List with status_code and response
}
\description{
Delete all external resources for a study
}
