% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{nada_admin_resource_import}
\alias{nada_admin_resource_import}
\title{Import an RDF file}
\usage{
nada_admin_resource_import(
  dataset_idno,
  rdf_file,
  skip_uploads = FALSE,
  overwrite = "no",
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{rdf_file}{RDF file path}

\item{skip_uploads}{TRUE/FALSE - If TRUE, won't upload files}

\item{overwrite}{yes/no - Overwrite existing resources}
}
\description{
Import an RDF file
}
