% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solr.R
\name{nada_admin_solr_index_entries}
\alias{nada_admin_solr_index_entries}
\title{SOLR}
\usage{
nada_admin_solr_index_entries(
  start_row = 0,
  limit = 10,
  loop = FALSE,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{start_row}{(required) row ID to start indexing from}

\item{limit}{Number of records to process}
}
\description{
Solr index catalog entries
}
