% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{nada_admin_study_attach_related}
\alias{nada_admin_study_attach_related}
\title{Attach related studies}
\usage{
nada_admin_study_attach_related(
  idno,
  related_datasets,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{(required) Study unique identifier}

\item{related_datasets}{\strong{(required)} list of related studies IDNOs}

\item{api_key}{API key (optional)}

\item{api_base_url}{API base endpoint (optional)}
}
\description{
Attach related studies
}
\examples{

nada_admin_study_attach_related (
  idno="survey-idno-test",
  related_datasets = c("idno-1", "idno-2", "idno-3")
)

}
