% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{nada_admin_study_delete}
\alias{nada_admin_study_delete}
\title{Delete a study}
\usage{
nada_admin_study_delete(idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(Required) Dataset IDNo}
}
\value{
list
}
\description{
Delete a single entry from the catalog
}
