% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{nada_admin_study_get_collections}
\alias{nada_admin_study_get_collections}
\title{Get owner and linked nada_collection_list by studies}
\usage{
nada_admin_study_get_collections(
  study_idno = NULL,
  offset = 0,
  limit = 1000,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{study_idno}{(Optional) Study/dataset IDNo}

\item{offset}{(Optional) Offset for pagination}

\item{limit}{(Optional) Number of rows to return}
}
\description{
Get owner and linked nada_collection_list by studies
}
