% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R
\name{nada_admin_study_get_json}
\alias{nada_admin_study_get_json}
\title{Get study metadata as JSON}
\usage{
nada_admin_study_get_json(
  idno,
  is_legacy = FALSE,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{Dataset IDNo}

\item{is_legacy}{TRUE | FALSE - if using NADA < 5.3, use legacy as TRUE}
}
\value{
Metadata as JSON
}
\description{
Get study metadata as JSON
}
