% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R
\name{nada_admin_study_replace_idno}
\alias{nada_admin_study_replace_idno}
\title{Replace study IDNO}
\usage{
nada_admin_study_replace_idno(
  old_idno,
  new_idno,
  api_key = NULL,
  api_base_url = NULL
)
}
\value{
list
}
\description{
Replace Study IDNO
}
