% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{nada_admin_widget_create}
\alias{nada_admin_widget_create}
\title{Create widget}
\usage{
nada_admin_widget_create(
  uuid,
  options = list(),
  zip_file,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{uuid}{\strong{(required)} Unique identifier for widget}

\item{options}{(list) widget options (title, description, thumbnail)}

\item{zip_file}{Zip file}

\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}
}
\value{
List with status_code and response
}
\description{
Create widget
}
