% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{nada_admin_widget_delete}
\alias{nada_admin_widget_delete}
\title{Delete widget}
\usage{
nada_admin_widget_delete(uuid, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{uuid}{(required) Widget ID}
}
\description{
Delete widget
}
