% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{nada_admin_widget_detach}
\alias{nada_admin_widget_detach}
\title{Detach widget from a study}
\usage{
nada_admin_widget_detach(idno, uuid, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(required) Study IDNo}

\item{uuid}{(required) Widget ID}
}
\description{
Detach widgetfrom study
}
