% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_code.R
\name{nada_country_codes}
\alias{nada_country_codes}
\title{Get Country Codes from World Bank Microdata API}
\usage{
nada_country_codes(
  api_key = NULL,
  api_base_url = NULL,
  name = NULL,
  iso = NULL,
  as_data_table = TRUE
)
}
\arguments{
\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}

\item{name}{Optional. Filter by country name.}

\item{iso}{Optional. Filter by ISO country code.}

\item{as_data_table}{Logical. If `TRUE` and format is JSON, converts the result to a `data.table`. Defaults to `TRUE`.}
}
\value{
A `data.table` if `as_data_table = TRUE`, otherwise a list (parsed JSON) or raw text for other formats.
}
\description{
Fetches country codes and related metadata from the World Bank Microdata Library API.
}
\examples{
\dontrun{
  dt <- nada_country_codes()
  print(dt)
}
}
