% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{nada_datafile_write_csv}
\alias{nada_datafile_write_csv}
\title{Export file to CSV}
\usage{
nada_datafile_write_csv(filepath = "", type = "", csvPath = "")
}
\arguments{
\item{filepath}{Path to data file}

\item{type}{Type of data file (dta | sav)}

\item{csvFile}{Output file path}
}
\value{
A list contains the status
}
\description{
Export datafiles (SPSS, STATA) to CSV format
}
\examples{
 nada_datafile_write_csv('/Data/example.dta', 'dta','/Data/example.csv')

}
