% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_get_api_key}
\alias{nada_get_api_key}
\title{Get API key}
\usage{
nada_get_api_key()
}
\value{
API key generated in NADA
}
\description{
`nada_get_api_key` returns the API key from global environment variables.
}
