% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_get_api_url}
\alias{nada_get_api_url}
\title{Get API URL}
\usage{
nada_get_api_url(endpoint = NULL)
}
\arguments{
\item{endpoint}{Endpoint to be added to the base endpoint URL, e.g., ""}
}
\value{
API base URL with optional endpoint appended
}
\description{
`nada_get_api_url` returns the API base endpoint URL from global environment variables.
}
