% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_get_api_verbose}
\alias{nada_get_api_verbose}
\title{Get API verbose mode}
\usage{
nada_get_api_verbose()
}
\value{
verbose parameter value
}
\description{
`nada_get_api_verbose` returns the API verbose parameter from the global environment variables.
}
