% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_get_verbose}
\alias{nada_get_verbose}
\title{Get verbose mode (alias for nada_get_api_verbose)}
\usage{
nada_get_verbose()
}
\value{
verbose parameter value
}
\description{
`nada_get_verbose` returns the API verbose parameter from the global environment variables.
This is an alias for `nada_get_api_verbose()`.
}
