% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nada_is_valid_url}
\alias{nada_is_valid_url}
\title{Check string is a URL}
\usage{
nada_is_valid_url(url)
}
\arguments{
\item{url}{URL string to validate.}
}
\value{
TRUE or FALSE
}
\description{
Check string is a URL
}
