% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nada_data_info.R
\name{nada_list_datasets}
\alias{nada_list_datasets}
\alias{nada_data_file_info}
\alias{nada_variables_info}
\alias{nada_list_variables}
\alias{nada_find_variable}
\title{Data File Metadata}
\usage{
nada_list_datasets(
  idno = NULL,
  api_key = NULL,
  api_base_url = NULL,
  as_data_table = TRUE
)

nada_data_file_info(
  idno = NULL,
  fileid = NULL,
  api_key = NULL,
  api_base_url = NULL,
  as_data_table = TRUE
)

nada_variables_info(
  idno = NULL,
  fileid = NULL,
  api_key = NULL,
  api_base_url = NULL,
  as_data_table = TRUE
)

nada_list_variables(
  idno = NULL,
  api_key = NULL,
  api_base_url = NULL,
  as_data_table = TRUE
)

nada_find_variable(
  idno = NULL,
  varId = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{Character. Study unique ID number.}

\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}

\item{as_data_table}{Logical. If `TRUE`, converts the result to `data.table`. Defaults to `TRUE`.}

\item{fileid}{Character. Data file unique ID number.}

\item{varId}{Character. Variable ID number.}
}
\value{
A `data.table`, a list (parsed JSON), or raw text, depending on `format` and `as_data_table`.
}
\description{
Functions to interact with data files and variables
}
\section{Functions}{
\itemize{
\item \code{nada_list_datasets()}: List all data files for a given study (`idno`).

\item \code{nada_data_file_info()}: Get information about a specific data file (`fileid`) in a study.

\item \code{nada_variables_info()}: Get variables in a specific data file (`fileid`) of a study.

\item \code{nada_list_variables()}: List all variables in a study (`idno`), regardless of data file.

\item \code{nada_find_variable()}: Get metadata for a single variable (`varId`) in a study.

}}
\examples{
\dontrun{
  dt <- nada_list_datasets(idno = "AFG_2015_DHS_v01_M")
}
\dontrun{
  dt <- nada_data_file_info(idno = "AFG_2015_DHS_v01_M",
  fileid = "F1")
  print(dt)
}
\dontrun{
  dt <- nada_variables_info(idno = "AFG_2015_DHS_v01_M",
  fileid = "F1")
  print(dt)
}
\dontrun{
  dt <- nada_list_variables(idno = "AFG_2015_DHS_v01_M")
  print(dt)
}
\dontrun{
  nada_find_variable(idno = "AFG_2015_DHS_v01_M", varId = "V2233")
}
}
