% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{nada_rdf_to_list}
\alias{nada_rdf_to_list}
\title{Convert RDF/XML to List}
\usage{
nada_rdf_to_list(rdf_file)
}
\arguments{
\item{rdf_file}{Path to RDF xml file}
}
\value{
List of resource metadata extracted from RDF file
}
\description{
Convert RDF/XML to list
}
