% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_set_api}
\alias{nada_set_api}
\title{Set API settings}
\usage{
nada_set_api(api_url, api_key, verbose = FALSE)
}
\arguments{
\item{api_url}{API base endpoint (URL)}

\item{api_key}{API key}

\item{verbose}{Verbose setting for API call (default is FALSE)}
}
\description{
`nada_set_api` is used to set the API settings consisting of the API URL, the API key and the verbose option.
`nada_set_api` is a wrapper for the functions `nada_set_api_url`, `nada_set_api_key` and `nada_set_api_verbose`.
}
\examples{
#nada_set_api("http://mynadacatalog.myorganization.org/index.php/api/", "abc123", verbose = TRUE)
}
