% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_set_api_key}
\alias{nada_set_api_key}
\title{Set API key}
\usage{
nada_set_api_key(api_key)
}
\arguments{
\item{api_key}{API Key}
}
\description{
`nada_set_api_key` is used to set the API key in the global environment variables.
}
