% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_set_api_url}
\alias{nada_set_api_url}
\title{Set API URL}
\usage{
nada_set_api_url(api_url)
}
\arguments{
\item{api_url}{API base endpoint}
}
\description{
`nada_set_api_url` is used to set the API base endpoint URL in the global environment variables.
}
