% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{nada_set_api_verbose}
\alias{nada_set_api_verbose}
\title{Set API verbose mode}
\usage{
nada_set_api_verbose(verbose = FALSE)
}
\arguments{
\item{verbose}{Verbose output to TRUE or FALSE}
}
\description{
Set API calls verbose options
}
