% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nada_data_access.R
\name{nada_study_access_types}
\alias{nada_study_access_types}
\title{Get list of data access types for studies}
\usage{
nada_study_access_types(
  as_data_table = TRUE,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{as_data_table}{Logical. If `TRUE`, converts the result to a `data.table`. Defaults to `TRUE`.}
}
\value{
A `data.table` if `as_data_table = TRUE`, otherwise a list (parsed JSON) or raw text for other formats.
}
\description{
Fetches the list of data access types available
}
\examples{
\dontrun{
  dt <- nada_study_access_types()
  print(dt)
}
}
