% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R, R/datasets.R
\name{nada_study_get_by_idno}
\alias{nada_study_get_by_idno}
\title{Find a study by IDNO}
\usage{
nada_study_get_by_idno(idno, api_key = NULL, api_base_url = NULL)

nada_study_get_by_idno(idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(required) Study unique identifier}

\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}
}
\value{
list
}
\description{
Find study by IDNO

Find a project by IDNO
}
\examples{

nada_study_get_by_idno (
  idno="survey-idno-test"
)

}
