% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R
\name{nada_study_information}
\alias{nada_study_information}
\title{Get study information excluding data dictionary}
\usage{
nada_study_information(idno = NULL, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{Character. Study unique ID number}

\item{api_key}{API key}

\item{api_base_url}{Base URL for the API}
}
\value{
A parsed JSON
}
\description{
Fetches detail metadata of a study without the excluding
}
\examples{
\dontrun{
  study <- nada_study_information(idno = "AFG_2015_DHS_v01_M")
  print(study)
}
}
