% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{nada_widget_list}
\alias{nada_widget_list}
\title{List all widgets}
\usage{
nada_widget_list(api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}
}
\value{
List of widgets
}
\description{
List all widgets
}
