% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{rename_collection}
\alias{rename_collection}
\title{Rename a collection}
\usage{
rename_collection(old_collectionid, new_collectionid, api_key = NULL,
  api_base_url = NULL)
}
\arguments{
\item{old_collectionid}{Old collection id}

\item{new_collectionid}{New collection id}
}
\description{
Change collection id of collection
}
