% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{resource_create}
\alias{resource_create}
\title{Create new study}
\usage{
resource_create(idno, dctype, title, dcformat = NULL, author = NULL,
  dcdate = NULL, country = NULL, language = NULL,
  contributor = NULL, publisher = NULL, rights = NULL,
  description = NULL, abstract = NULL, toc = NULL,
  file_path = NULL, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(required) Study unique identifier}

\item{type}{(required) Type of study - survey, geospatial, table, document, timeseries}

\item{repositoryid}{Collection ID that owns the study}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}

\item{metadata}{\strong{(required)} Metadata list depending on the type of study}
}
\description{
Create a new study
}
\examples{


doc_desc=list(
"idno"="doc-idno",
"producers"=list(
list(
    "name"="name here",
    "abbr"="abbreviation"
  )
)
)

study_desc=list(
  "title_statement"= list(
    "idno"= "survey-idno-test",
    "title"= "string",
    "sub_title"= "string",
    "alternate_title"= "string",
    "translated_title"= "string"
  ),
  "study_info"=list(
    "nation"=list(
      list(
        "name"="Test",
        "abbreviation"="tst")
    )
  )
)

create (
  idno="survey-idno-test",
  type="survey",
  published = 1,
  overwrite = "yes",
  doc_desc = doc_desc,
  study_desc = study_desc,
  data_files=NULL,
  variables=NULL,
  variable_groups = NULL,
  additional = NULL
)




}
