% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{resource_download}
\alias{resource_download}
\title{Download resource file}
\usage{
resource_download(
  api_key = NULL,
  api_base_url = NULL,
  dataset_idno,
  resource_id
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{resource_id}{Resource ID}
}
\value{
file
}
\description{
Download resource file
}
