% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{resource_upload}
\alias{resource_upload}
\title{Upload external resources}
\usage{
resource_upload(api_key = NULL, api_base_url = NULL, dataset_idno,
  resource_id = NULL, file)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{resource_id}{(Optional) External resource ID}

\item{file}{External resource file to be uploaded}
}
\description{
Upload an external resource file
}
