% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{resources}
\alias{resources}
\title{List External Resources}
\usage{
resources(api_key = NULL, api_base_url = NULL, dataset_idno)
}
\arguments{
\item{dataset_idno}{Study IDNo}
}
\value{
List of external resources
}
\description{
List external resources for a study
}
