% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R, R/globals.R
\name{set_api_url}
\alias{set_api_url}
\title{Set API URL}
\usage{
set_api_url(api_url)

set_api_url(api_url)
}
\arguments{
\item{api_url}{API base endpoint}
}
\description{
`api_url` is used to set the API base endpoint URL in the global environment variables.

Set API Base URL
}
