% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R, R/globals.R
\name{set_api_verbose}
\alias{set_api_verbose}
\title{set_verbose}
\usage{
set_api_verbose(verbose = FALSE)

set_api_verbose(verbose = FALSE)
}
\arguments{
\item{verbose}{Verbose output to TRUE or FALSE}
}
\description{
Set API calls verbose options

Set API calls verbose options
}
