% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_api.R
\name{set_editor_api_key}
\alias{set_editor_api_key}
\title{Set API key}
\usage{
set_editor_api_key(api_key)
}
\arguments{
\item{api_key}{API Key}
}
\description{
`set_editor_api_key` is used to set the API key in the global environment variables.
}
