% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_api.R
\name{set_editor_api_url}
\alias{set_editor_api_url}
\title{Set Editor API URL}
\usage{
set_editor_api_url(api_url)
}
\arguments{
\item{api_url}{API base endpoint}
}
\description{
`editor_api_url` is used to set the API base endpoint URL in the global environment variables.
}
