% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_api.R
\name{set_editor_api_verbose}
\alias{set_editor_api_verbose}
\title{set_editor_verbose}
\usage{
set_editor_api_verbose(verbose = FALSE)
}
\arguments{
\item{verbose}{Verbose output to TRUE or FALSE}
}
\description{
Set API calls verbose options
}
