% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_compability.R
\name{set_verbose}
\alias{set_verbose}
\title{(deprecated - use set_api_verbose) Set API verbose option}
\usage{
set_verbose(verbose = FALSE)
}
\arguments{
\item{verbose}{Verbose output to TRUE or FALSE}
}
\description{
Deprecated function, available for backward compability
}
