% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{upload_thumbnail}
\alias{upload_thumbnail}
\title{Upload thumbnail for a study}
\usage{
upload_thumbnail(idno, thumbnail, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(required) Study unique identifier}

\item{thumbnail}{\strong{(required)} Path to the thumbnail file}
}
\description{
Upload thumbnail for a study
}
\examples{

upload_thumbnail (
  idno="survey-idno-test",
  thumbnail = "/thumbnails/thumbnail-idno-test.png"
)

}
