% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Videos.R
\name{video_add}
\alias{video_add}
\title{Create new video}
\usage{
video_add(
  idno,
  metadata,
  repositoryid = NULL,
  access_policy = NULL,
  data_remote_url = NULL,
  published = 0,
  overwrite = "no",
  thumbnail = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{(required) Unique identifier}

\item{metadata}{\strong{(required)} Metadata using video schema}

\item{repositoryid}{Owner Collection ID}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}
}
\description{
Create a new video
}
