% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{widgets_attach}
\alias{widgets_attach}
\title{Attach widget to a study}
\usage{
widgets_attach(idno, uuid, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{(required) Study IDNo}

\item{uuid}{(required) Widget ID}
}
\description{
Attach widget to a study
}
