% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{widgets_create}
\alias{widgets_create}
\title{Create widget}
\usage{
widgets_create(
  uuid,
  options = list(),
  zip_file,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{uuid}{\strong{(required)} Unique identifier for widget}

\item{options}{(list) widget options (title, description, thumbnail)}

\item{zip_file}{Zip file}
}
\value{

}
\description{
Create widget
}
