% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{widgets_delete}
\alias{widgets_delete}
\title{Delete widget}
\usage{
widgets_delete(uuid, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{uuid}{(required) Widget ID}
}
\description{
Delete widget
}
