% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_entries.R
\name{write_study_json}
\alias{write_study_json}
\title{Write study metadata as JSON}
\usage{
write_study_json(
  idno,
  output_file,
  is_legacy = FALSE,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{Dataset IDNo}

\item{output_file}{Path to the output file}

\item{is_legacy}{TRUE | FALSE - if using NADA < 5.3, use legacy as TRUE}
}
\value{
None
}
\description{
Write study metadata as JSON
}
