% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_pdf_cover}
\alias{capture_pdf_cover}
\title{Capture cover page of a PDF}
\usage{
capture_pdf_cover(
  file_path,
  file_name_jpg = NULL,
  file_dir_jpg = NULL,
  dpi = 72
)
}
\arguments{
\item{file_path}{\strong{(required)} PDF file path}

\item{file_name_jpg}{\strong{(optional)} file name to store jpg. If empty, file name of jpg is same as pdf}

\item{file_dir_jpg}{\strong{(optional)} file directory to store jpg. If empty, file will be stored in same folder as pdf}

\item{dpi}{set image resolution in dpi, default is 72 dpi}
}
\value{
image file
}
\description{
Capture cover page of a PDF file
}
