% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{dataset_collections_list}
\alias{dataset_collections_list}
\title{Get owner and linked collections by studies}
\usage{
dataset_collections_list(study_idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{study_idno}{(Optional) Study/dataset IDNo}
}
\description{
Get owner and linked collections by studies
}
