% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{GetDatasets}
\usage{
datasets(
  idno = NULL,
  offset = 0,
  limit = 50,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{(Optional) Dataset IDNo}

\item{offset}{Specify the number of rows to skip, default is 0}

\item{limit}{Specify number of rows to return, default is 50 rows. Note: if more than 500 rows are requested, several API calls are made and results are combined..}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\value{
List of studies or a single study info
}
\description{
Load a list of all datasets or get info for a single dataset/study
}
