% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nada_admin_data_table_upload_csv}
\alias{nada_admin_data_table_upload_csv}
\title{Upload CSV or ZIP File}
\usage{
nada_admin_data_table_upload_csv(
  db_id,
  table_id,
  file,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{db_id}{(Required) database name}

\item{table_id}{(Required) Table name}

\item{file}{(Required) CSV or ZIP file path}

\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}
}
\description{
nada_admin_data_table_upload_csv
}
\details{
Upload a CSV or ZIP file
}
\examples{

nada_admin_data_table_upload_csv (
  file = "path/to/data.csv"
)

nada_admin_data_table_upload_csv (
  file = "path/to/data.zip"
)

}
