% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{nada_admin_study_import_ddi}
\alias{nada_admin_study_import_ddi}
\title{ImportDDI}
\usage{
nada_admin_study_import_ddi(
  api_key = NULL,
  api_base_url = NULL,
  xml_file = NULL,
  rdf_file = NULL,
  repositoryid = NULL,
  overwrite = "no",
  access_policy = NULL,
  data_remote_url = NULL,
  published = NULL
)
}
\arguments{
\item{api_key}{API key (optional if API key is set using nada_set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using nada_set_api_url)}

\item{xml_file}{(Required) DDI/XML file path}

\item{rdf_file}{RDF file path}

\item{repositoryid}{Collection ID that owns the study}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{verbose}{Show verbose output - True, False}
}
\description{
Import a DDI file
}
