% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{nada_collection_list}
\alias{nada_collection_list}
\title{Get all collections}
\usage{
nada_collection_list(repositoryid = NULL, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{repositoryid}{(Optional) Collection IDNo}
}
\value{
List of studies or a single study info
}
\description{
Load a list of all nada_collection_list or get info for a single collection
}
