% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nada_http_delete}
\alias{nada_http_delete}
\title{Make an http DELETE request}
\usage{
nada_http_delete(url, options = list(), api_key = NULL, api_base_url = NULL)
}
\value{
http request output
}
\description{
Make an http DELETE request
}
