% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nada_http_put}
\alias{nada_http_put}
\title{Make an http PUT request}
\usage{
nada_http_put(
  url,
  options = list(),
  request_encode = "json",
  api_key = NULL,
  api_base_url = NULL
)
}
\value{
http request output


request_encode = c("multipart", "form", "json", "raw")
}
\description{
Make an http PUT request
}
