% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{nada_resource_download}
\alias{nada_resource_download}
\title{Download resource file}
\usage{
nada_resource_download(
  dataset_idno,
  resource_id,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{dataset_idno}{Study IDNo}

\item{resource_id}{Resource ID}
}
\value{
file
}
\description{
Download resource file
}
