% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R, R/globals.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key}
\usage{
set_api_key(api_key)

set_api_key(api_key)
}
\arguments{
\item{api_key}{API Key}
}
\value{
A matrix of the infile
}
\description{
`set_api_key` is used to set the API key in the global environment variables.

Set API key
}
