% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Videos.R
\name{video_add}
\alias{video_add}
\title{Create new video}
\usage{
video_add(
  idno,
  metadata,
  repositoryid = NULL,
  access_policy = NULL,
  data_remote_url = NULL,
  published = 0,
  overwrite = "no",
  thumbnail = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{(required) Unique identifier}

\item{metadata}{\strong{(required)} Metadata using video schema}

\item{repositoryid}{Owner Collection ID}

\item{access_policy}{Select the access policy suitable for your data. Valid values - "open" "direct" "public" "licensed" "enclave" "remote" "other"}

\item{data_remote_url}{Link to the website where the data is available. Required if access_policy is set to 'remote'.}

\item{published}{Set status for study - 0 = Draft, 1 = Published}

\item{overwrite}{Overwrite if a study with the same ID already exists? Valid values "yes", "no"}
}
\description{
Create a new video
}
\examples{

metadata=list(
   "repositoryid"= "central",
   "published"= "1",
   "overwrite"= "no",
   "video_description"= list(
     "idno"= "video-idno",
     "title"= "Lorem ipsum",
     "description"= "Description of video",
     "video_provider"= "youtube",
     "video_url"= "https://www.youtube.com/watch?v=7X8II6J-6mU",
     "embed_url"= "https://www.youtube.com/embed/7X8II6J-6mU",
     "country"= "Country",
     "language"= "english",
     "contributor"= "Contributor name",
     "publisher"= "Publisher",
     "rights"= "Rights",
     "author"= "Author name",
     "date_created"= "2020-01-01",
     "date_published"= "2021-08-17"
   ),
   "tags"= list(
     list(
       "tag"= "demo"
     )
   )
)


 video_add (
  idno="video-idno",
  published = 1,
  overwrite = "yes",
  metadata = metadata,
  thumbnail ="path-to-thumbnail-file.jpg"
)




}
